+++ NEWS.md
@@ -1,9 +1,10 @@
 # dirichletprocess 0.2.2.900
 
-* Fixed bug in PosteriorClusters and PosteriorFunction
+* Fixed bug in PosteriorClusters and PosteriorFunction.
 * Added in new Beta mixture model for avoiding boundary. 
 * Fixed a bug in ChangeObservations when using more than one dimension. 
-* Added in Burn, Print and Diagnostic Plots 
+* Added in Burn, Print and Diagnostic Plots.
+* Reworked the PriorParametersUpdate function to be consistent. 
 
 # dirichletprocess 0.2.2